update master branch

+++ app/Http/Controllers/ChecklistTemplateController.php
@@ -4,56 +4,47 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\SearchRequest;
 use App\Http\Resources\ChecklistTemplateIndexResource;
-use App\Models\Checklist;
+use Artwork\Modules\Checklist\Models\Checklist;
 use App\Models\ChecklistTemplate;
-use App\Models\Department;
 use App\Models\TaskTemplate;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Redirect;
+use Inertia\Response;
+use Inertia\ResponseFactory;
 
 class ChecklistTemplateController extends Controller
 {
-
     public function __construct()
     {
         $this->authorizeResource(ChecklistTemplate::class);
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Inertia\Response|\Inertia\ResponseFactory
-     */
-    public function index()
+    public function index(): Response|ResponseFactory
     {
         return inertia('ChecklistTemplates/ChecklistTemplateManagement', [
             'checklist_templates' => ChecklistTemplateIndexResource::collection(ChecklistTemplate::all())->resolve(),
         ]);
     }
 
-    public function search(SearchRequest $request)
+    /**
+     * @param SearchRequest $request
+     * @return array<string, mixed>
+     */
+    public function search(SearchRequest $request): array
     {
-        return ChecklistTemplateIndexResource::collection(ChecklistTemplate::search($request->input('query')))
-            ->resolve();
+        return ChecklistTemplateIndexResource::collection(
+            ChecklistTemplate::search($request->input('query'))->get()
+        )->resolve();
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Inertia\Response|\Inertia\ResponseFactory
-     */
-    public function create()
+    public function create(): Response|ResponseFactory
     {
         return inertia('ChecklistTemplates/Create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function store(Request $request)
+    public function store(Request $request): RedirectResponse
     {
         if ($request->checklist_id) {
             $this->createFromChecklist($request);
@@ -61,10 +52,10 @@ class ChecklistTemplateController extends Controller
             $this->createFromScratch($request);
         }
 
-        return Redirect::route('checklist_templates.management')->with('success', 'ChecklistTemplate created.');
+        return Redirect::route('checklist_templates.management');
     }
 
-    protected function createFromChecklist(Request $request)
+    protected function createFromChecklist(Request $request): void
     {
         $checklist = Checklist::where('id', $request->checklist_id)->first();
 
@@ -77,110 +68,64 @@ class ChecklistTemplateController extends Controller
             TaskTemplate::create([
                 'name' => $task->name,
                 'description' => $task->description,
+                'done' => false,
                 'checklist_template_id' => $checklist_template->id
             ]);
         }
 
-        $checklist_template->departments()->sync(
-            collect($checklist->departments)
-                ->map(function ($department) {
-                    $this->authorize('update', Department::find($department['id']));
-
-                    return $department['id'];
-                })
-        );
+        $checklist_template->users()->sync(collect($checklist->users)->pluck('id'));
     }
 
-    protected function createFromScratch(Request $request)
+    protected function createFromScratch(Request $request): void
     {
         $checklist_template = ChecklistTemplate::create([
             'name' => $request->name,
             'user_id' => $request->user_id
         ]);
 
-        $checklist_template->departments()->sync(
-            collect($request->departments)
-                ->map(function ($department) {
-                    $this->authorize('update', Department::find($department['id']));
-
-                    return $department['id'];
-                })
-        );
+        $checklist_template->users()->sync(Collection::make($request->users)->pluck('id'));
 
         if ($request->task_templates) {
             $checklist_template->task_templates()->createMany($request->task_templates);
         }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\ChecklistTemplate  $checklistTemplate
-     * @return \Inertia\Response|\Inertia\ResponseFactory
-     */
-    public function show(ChecklistTemplate $checklistTemplate)
+    public function show(ChecklistTemplate $checklistTemplate): Response|ResponseFactory
     {
         return inertia('ChecklistTemplates/Show', [
             'checklist_template' => new ChecklistTemplateIndexResource($checklistTemplate)
         ]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\ChecklistTemplate  $checklistTemplate
-     * @return \Inertia\Response|\Inertia\ResponseFactory
-     */
-    public function edit(ChecklistTemplate $checklistTemplate)
+    public function edit(ChecklistTemplate $checklistTemplate): Response|ResponseFactory
     {
         return inertia('ChecklistTemplates/Edit', [
             'checklist_template' => new ChecklistTemplateIndexResource($checklistTemplate)
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\ChecklistTemplate  $checklistTemplate
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function update(Request $request, ChecklistTemplate $checklistTemplate)
+    public function update(Request $request, ChecklistTemplate $checklistTemplate): RedirectResponse
     {
         $checklistTemplate->update($request->only('name'));
 
-        $checklistTemplate->departments()->sync(
-            collect($request->departments)
-                ->map(function ($department) {
-                    $this->authorize('update', Department::find($department['id']));
+        $userIdsToSync = Collection::make($request->users)->pluck('id');
+        $checklistTemplate->users()->sync($userIdsToSync);
 
-                    return $department['id'];
-                })
-        );
-
-        $checklistTemplate->task_templates()->delete();
-
-        foreach ($request->task_templates as $task_template) {
-            if (! is_array($task_template)) {
-                $checklistTemplate->task_templates()->save($task_template);
-            } else {
-                $checklistTemplate->task_templates()->create($task_template);
+        if ($request->task_templates) {
+            $checklistTemplate->task_templates()->delete();
+            foreach ($request->task_templates as $task_template) {
+                $task_template_new = $checklistTemplate->task_templates()->create($task_template);
+                $task_template_new->task_users()->sync($userIdsToSync);
             }
         }
 
-        return Redirect::route('checklist_templates.management', $checklistTemplate->id)->with('success', 'ChecklistTemplate updated');
+        return Redirect::route('checklist_templates.management', $checklistTemplate->id);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\ChecklistTemplate  $checklistTemplate
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function destroy(ChecklistTemplate $checklistTemplate)
+    public function destroy(ChecklistTemplate $checklistTemplate): RedirectResponse
     {
         $checklistTemplate->delete();
 
-        return Redirect::back()->with('success', 'ChecklistTemplate deleted');
+        return Redirect::back();
     }
 }
